requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
dj-static==0.0.6
Django==3.2.20
gunicorn==20.1.0
line-bot-sdk==2.3.0
psycopg2==2.9.5
pytz==2022.2.1
qrcode==7.4.2
sqlparse==0.4.3
static3==0.7.0
typing_extensions==4.3.0
whitenoise==6.2.0
wincertstore==0.2
django-import-export == 3.0.2